+++ saleor/graphql/account/tests/mutations/staff/test_customer_create.py
@@ -85,7 +85,7 @@ CUSTOMER_CREATE_MUTATION = """
 
 
 @patch("saleor.plugins.manager.PluginsManager.customer_metadata_updated")
-@patch("saleor.account.notifications.default_token_generator.make_token")
+@patch("saleor.account.notifications.token_generator.make_token")
 @patch("saleor.plugins.manager.PluginsManager.notify")
 @patch("saleor.plugins.manager.PluginsManager.account_set_password_requested")
 def test_customer_create(
@@ -203,7 +203,7 @@ def test_customer_create(
 
 
 @patch("saleor.plugins.manager.PluginsManager.customer_metadata_updated")
-@patch("saleor.account.notifications.default_token_generator.make_token")
+@patch("saleor.account.notifications.token_generator.make_token")
 @patch("saleor.plugins.manager.PluginsManager.notify")
 @patch("saleor.plugins.manager.PluginsManager.account_set_password_requested")
 def test_customer_create_as_app(
@@ -325,7 +325,7 @@ def test_customer_create_as_app(
     )
 
 
-@patch("saleor.account.notifications.default_token_generator.make_token")
+@patch("saleor.account.notifications.token_generator.make_token")
 @patch("saleor.plugins.manager.PluginsManager.notify")
 def test_customer_create_send_password_with_url(
     mocked_notify,
@@ -513,7 +513,7 @@ def test_customer_create_with_non_unique_external_reference(
     assert error["message"] == "User with this External reference already exists."
 
 
-@patch("saleor.account.notifications.default_token_generator.make_token")
+@patch("saleor.account.notifications.token_generator.make_token")
 @patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_customer_create_webhook_event_triggered(
     mocked_trigger_webhooks_async,